trickya, roll it back
